Move the logic to a separate component
